Car_Localizer/Car_Localizer/requirements.txt
appdirs==1.4.4
attrs==21.4.0
black==19.10b0
Bottleneck==1.3.5
certifi==2022.6.15
cffi==1.15.0
cfgv==3.3.1
click==8.1.1
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.0.0
colorama==0.4.4
cycler==0.11.0
cytoolz==0.11.0
dask==2022.5.0
distlib==0.3.2
filelock==3.6.0
Fiona==1.8.13.post1
fonttools==4.25.0
fsspec==2022.3.0
GDAL==3.0.2
geopandas==0.9.0
graphviz==0.17
identify==2.5.2
imagecodecs==2021.8.26
imageio==2.9.0
joblib==1.1.0
kiwisolver==1.4.2
locket==1.0.0
mapclassify==2.4.3
matplotlib==3.5.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
munch==2.5.0
munkres==1.1.4
mypy-extensions==0.4.3
networkx==2.8.4
nodeenv==1.7.0
numexpr==2.8.3
numpy==1.22.3
opencv==4.6.0
packaging==21.3
pandas==1.4.3
partd==1.2.0
pathspec==0.9.0
Pillow==6.2.1
pip==22.1.2
platformdirs==2.5.1
plotly==5.9.0
pre-commit==2.20.0
pycparser==2.21
pyparsing==3.0.4
pyproj==2.6.1.post1
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
regex==2022.7.9
Rtree==0.9.7
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
setuptools==61.2.0
Shapely==1.7.1
sip==4.19.13
six==1.16.0
tenacity==8.0.1
threadpoolctl==2.2.0
tifffile==2021.7.2
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
torch==1.9.0
torchvision==0.13.1
torchviz==0.0.2
tornado==6.1
typed-ast==1.4.3
typing_extensions==4.1.1
ukkonen==1.0.1
virtualenv==20.4.6
wheel==0.37.1
wincertstore==0.2
